feat: Refactor menu and VFO handling in applications for improved functionality

+++ src/ui/ui.h
@@ -426,15 +426,23 @@ public:
        // Format the frequency string based on the precision and whether it's in kHz
         if (isKHz) {
             if (precision == 0) {
-                snprintf(uiBuffer, CHAR_BUFFER_SIZE, "%u.%03u KHz", frequency / 1000, (frequency % 1000) / 10);
+                snprintf(uiBuffer, CHAR_BUFFER_SIZE, "%u.%03u KHz",
+                         static_cast<unsigned int>(frequency / 1000),
+                         static_cast<unsigned int>((frequency % 1000) / 10));
             } else {
-                snprintf(uiBuffer, CHAR_BUFFER_SIZE, "%u.%03u KHz", frequency / 1000, frequency % 1000);
+                snprintf(uiBuffer, CHAR_BUFFER_SIZE, "%u.%03u KHz",
+                         static_cast<unsigned int>(frequency / 1000),
+                         static_cast<unsigned int>(frequency % 1000));
             }
         } else {
             if (precision == 0) {
-                snprintf(uiBuffer, CHAR_BUFFER_SIZE, "%u.%03u Hz", frequency / 1000, (frequency % 1000) / 10);
+                snprintf(uiBuffer, CHAR_BUFFER_SIZE, "%u.%03u Hz",
+                         static_cast<unsigned int>(frequency / 1000),
+                         static_cast<unsigned int>((frequency % 1000) / 10));
             } else {
-                snprintf(uiBuffer, CHAR_BUFFER_SIZE, "%u.%03u Hz", frequency / 1000, frequency % 1000);
+                snprintf(uiBuffer, CHAR_BUFFER_SIZE, "%u.%03u Hz",
+                         static_cast<unsigned int>(frequency / 1000),
+                         static_cast<unsigned int>(frequency % 1000));
             }
         }
         return uiBuffer;
